fix:增加啟動task的工作的api

+++ app/(home)/projects/[projectName]/applications/[applicationName]/dashboard/preprocessing_pipeline/[prePipeName]/tasks/service.jsx
@@ -60,10 +60,10 @@ export const useFetchTaskFile = (pipelineUID) => {
 
 //創建task
 export const useCreateTask = () => {
-  const createTask = async (formData) => {
-    if (formData) {
+  const createTask = async (createTaskData) => {
+    if (createTaskData) {
       //TaskMetadataWriter/create
-      const response = await getAPI("Za0lf5Tf5pI3fhMx", formData);
+      const response = await getAPI("Za0lf5Tf5pI3fhMx", createTaskData);
       if (response.status === 200) {
         return response.data;
       } else if (response && response instanceof Error) {
@@ -72,6 +72,22 @@ export const useCreateTask = () => {
     }
   };
   return { createTask };
+};
+
+//啟動preprocessing task
+export const useRunPreprocessingTask = () => {
+  const runTask = async (formData) => {
+    if (formData) {
+      //TaskWorker/preprocessing
+      const response = await getAPI("jPqyAFWh7hBKRRNK", formData);
+      if (response.status === 200) {
+        return response.data;
+      } else if (response && response instanceof Error) {
+        console.error("Error running task:", response.data);
+      }
+    }
+  };
+  return { runTask };
 };
 
 //更新task
@@ -90,7 +106,7 @@ export const useUpdateTask = (formData) => {
   return { updateTask };
 };
 
-//刪除config
+//刪除task
 export const useDeleteTask = (taskUID) => {
   const deleteTask = async () => {
     if (taskUID) {
@@ -107,6 +123,23 @@ export const useDeleteTask = (taskUID) => {
   return { deleteTask };
 };
 
+//刪除task的工作
+export const useDeleteTaskWorker = (taskUID) => {
+  const deleteTaskWorker = async () => {
+    if (taskUID) {
+      //TaskWorker/delete
+      const data = { task_uid: taskUID };
+      const response = await getAPI("CNNfUmDpcWXp7vdM", data);
+      if (response.status === 200) {
+        return response.data;
+      } else if (response && response instanceof Error) {
+        console.error("Error deleting task", response.data);
+      }
+    }
+  };
+  return { deleteTaskWorker };
+};
+
 export const HandleUpdate = async (formData, onEdit, onClose) => {
   const { updateTask } = useUpdateTask(formData);
   const response = await updateTask();
@@ -116,21 +149,44 @@ export const HandleUpdate = async (formData, onEdit, onClose) => {
   }
 };
 
-export const HandleDelete = async (configUID, onDelete, onClose) => {
-  const { deleteTask } = useDeleteTask(configUID);
+export const HandleDelete = async (taskUID, onDelete, onClose) => {
+  const { deleteTask } = useDeleteTask(taskUID);
+  const { deleteTaskWorker } = useDeleteTaskWorker(taskUID);
+
   const response = await deleteTask();
   if (response && !(response instanceof Error)) {
-    onDelete();
-    onClose();
+    //如果metadata刪除成功，呼叫deleteTaskWorker動作
+    const deleteTaskWorkerResponse = await deleteTaskWorker();i
+    if (
+      deleteTaskWorkerResponse &&
+      !(deleteTaskWorkerResponse instanceof Error)
+    ) {
+      onDelete();
+      onClose();
+    }
   }
 };
 
 export const HandleCreate = async (formData, onCreate, onClose) => {
   const { createTask } = useCreateTask();
-  const response = await createTask(formData);
+  const { runTask } = useRunPreprocessingTask();
+
+  //傳遞到createTask api所需的資料
+  const createTaskData = {
+    name: formData.task_name,
+    description: formData.task_description,
+    f_pipeline_uid: formData.pipeline_uid,
+  };
+  const response = await createTask(createTaskData);
   if (response && !(response instanceof Error)) {
-    onCreate();
-    onClose();
+    // 如果任務創建成功，接著呼叫 runTask 來啟動任務
+    const runTaskResponse = await runTask(formData);
+
+    if (runTaskResponse && !(runTaskResponse instanceof Error)) {
+      // 執行成功後觸發 onCreate 和 onClose
+      onCreate();
+      onClose();
+    }
   }
 };
